requirements.txt
asgiref==3.6.0
Django==4.2
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
